docs: "Jelly Slider" code refactor
