+++ apps/typegpu-docs/tests/individual-example-tests/utils/commonMocks.ts
@@ -10,13 +10,15 @@ export function setupCommonMocks() {
         'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/146.0.0.0 Safari/537.36',
     });
 
-    vi.spyOn(HTMLCanvasElement.prototype, 'getContext').mockImplementation(() => {
-      return createDeepNoopProxy(
-        {} as unknown as CanvasRenderingContext2D,
-        new Set(),
-        // oxlint-disable-next-line typescript/no-explicit-any -- we testing here
-      ) as any;
-    });
+    vi.spyOn(HTMLCanvasElement.prototype, 'getContext').mockImplementation(
+      function (this: HTMLCanvasElement) {
+        return createDeepNoopProxy(
+          { canvas: this } as unknown as CanvasRenderingContext2D,
+          new Set(),
+          // oxlint-disable-next-line typescript/no-explicit-any -- we testing here
+        ) as any;
+      },
+    );
 
     Object.defineProperty(HTMLCanvasElement.prototype, 'width', {
       get: () => 256,
